Refactoring is finished
