+++ rs-marathon-st/src/components/CardsList/style.scss
